split onehot opl
