+++ onnx/src/ops/array/one_hot.rs
@@ -1,5 +1,4 @@
 use tract_hir::internal::*;
-use tract_ndarray::Dimension;
 
 use crate::model::ParsingContext;
 use crate::pb::NodeProto;
@@ -52,7 +51,7 @@ impl Expansion for OneHot {
                 bail!("Expected positive dimension, got {}", dim)
             }
             let (off, on) = dispatch_datum!(Self::split_values_t(values.datum_type())(&values))?;
-            let op = MirOneHot { axis, dim: dim as usize, off, on };
+            let op = tract_onnx_opl::one_hot::OneHot { axis, dim: dim as usize, off, on };
             model.wire_node(prefix, op, &[inputs[0]])
         } else {
             bail!("Expected dim and value to be determined, got {:?} and {:?}", dim, values)
@@ -84,67 +83,5 @@ impl Expansion for OneHot {
                 s.equals(&outputs[0].shape[axis], dim.to_dim())
             })
         })
-    }
-}
-
-#[derive(Debug, PartialEq, Clone, new, Hash)]
-struct MirOneHot {
-    axis: usize,
-    dim: usize,
-    off: Tensor,
-    on: Tensor,
-}
-
-tract_linalg::impl_dyn_hash!(MirOneHot);
-
-impl Op for MirOneHot {
-    fn name(&self) -> Cow<str> {
-        "MirOnehot".into()
-    }
-
-    op_onnx!();
-    op_as_typed_op!();
-}
-
-impl TypedOp for MirOneHot {
-    fn output_facts(&self, inputs: &[&TypedFact]) -> TractResult<TVec<TypedFact>> {
-        let mut shape = inputs[0].shape.to_tvec();
-        shape.insert(self.axis, self.dim.to_dim());
-        Ok(tvec!(TypedFact::dt_shape(self.off.datum_type(), &*shape)?))
-    }
-
-    as_op!();
-}
-
-impl EvalOp for MirOneHot {
-    fn is_stateless(&self) -> bool {
-        true
-    }
-
-    fn eval(&self, mut inputs: TVec<Arc<Tensor>>) -> TractResult<TVec<Arc<Tensor>>> {
-        let input = args_1!(inputs);
-        let output = dispatch_datum!(Self::eval_t(self.off.datum_type())(self, &input))?;
-        Ok(tvec!(output.into_arc_tensor()))
-    }
-}
-
-impl MirOneHot {
-    fn eval_t<T: Datum + Clone>(&self, input: &Tensor) -> TractResult<Tensor> {
-        let off = self.off.to_scalar::<T>()?;
-        let on = self.on.to_scalar::<T>()?;
-        let mut shape: TVec<usize> = input.shape().into();
-        shape.insert(self.axis, self.dim);
-        let mut array = tract_ndarray::ArrayD::<T>::from_elem(&*shape, off.to_owned());
-        let input = input.cast_to::<i32>()?;
-        let input = input.to_array_view::<i32>()?;
-        dbg!(&input);
-        for icoord in tract_ndarray::indices_of(&input) {
-            let mut ocoord:Vec<usize> = icoord.slice().into();
-            let coord = input[&icoord];
-            let coord = if coord < 0 { coord + self.dim as i32 } else { coord } as usize;
-            ocoord.insert(self.axis, coord);
-            array[&*ocoord] = on.clone();
-        }
-        Ok(array.into_tensor())
     }
 }
